Slowly moving toward better map structure
Instead of taking the raw array from the object of the generator class, we now
just store the object itself as the map.

Also, items added to the map (e.g. stairwells, inventory, etc.) are store as
objects with x/y positions, rather than placed on the map.  have a proof of
concept running with the stairs.

Near future improvements:

 - create a div that is overlaid on the map, but below the player, where items
   can be placed without affecting the map content
 - expand the map structure to work recursively, ultimately having a top-level
   map that is the world, and children will be attached thereunder.
 - stairs will need to be more complicated objects, storing info such as where
   they go, if their target still needs to be generated, etc.
 - similarily, worlds will need to store data to keep game info consistent
   (items being collected, user location, etc.)