Improved kbListener lib and updated game to use it
kbListener was previously a few generically named functions with a limited
keyboard map.  It has now been grouped together into a kbListener class.

The mapping is now done more dynamically to match different browsers..
different keyboards will be another matter.

Also, event catching has been improved.  Instead of simply assigning a callback
to the keyup or keydown events, it instead checks for an existing callback and
wraps them in a function.

Also modified the game to work with it.