Map loading and sprite fixes, added water sprite
Fixed one issue with map loading, the canvas needed its size specified before
loading the image and I missed that step, hench much of the world not loading.

There were issues with canvas spritesets due to a recent "improvement" I did,
which had a set of default data to assign.  That set included some objects and
arrays, which ended up with templates using the same ones.  As a result,
overlapping frame names caused issues.

Added a sprite for water.

Made the map character set more dynamic, allowing it to be specified by JSON
info for map loading.

Added a shrunken version of the new map (200x200) as the 1000x1000 one takes
too long to load and seems to be pushing limits.