Moving toward including PathFinding.js
Very close to getting it to work.  I have the map generating the input I want,
the next step is to get it find a path on that map.  As-is, the remarked code
gets an error when calling the function to plot the path - I don't know why.

Once that's solved, using that multi-step path should be easy.  It's just a
matter of assigning the steps in that path as targets for the character.  those
already get shifted off an array, and should pretty much work right away.