Several updates
Removed the path finding code that I was using, and downloaded a new library
that seems more straightforward and compatible with the available code.  Not
yet implemented, but ready to go.

Moved the player location control out of the map and into the player.  Maps
still set location when the player enters the world, but they are not used to
track the location (as that limits the tracking to the active player and no
other entities.  Now the player has both a map position and a pixel position.

Added an in-page debug console, replacing the console.log function with one
that writes to that one instead.  Just because, but it does allow debug output
when in full screen mode.

Added a favicon